itmo_master_advisor/requirements.txt
python-telegram-bot==20.7
openai==1.12.0
chromadb==0.4.22
langchain==0.1.6
langchain-openai==0.0.5
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.17.2
webdriver-manager==4.0.1
python-dotenv==1.0.1
pydantic==2.6.1
aiohttp==3.9.3